requirements.txt
progressbar2
opencv-python
matplotlib
tensorflow
keras